Fefactor using modules to make it more readable
