+++ index.html
@@ -58,9 +58,10 @@
   </section>
 
   <footer class="copyright">
-    <p>Copyright</p>
+    <p>Copyright by Santiago and Andrés</p>
   </footer>
 
-  <script src="./index.js"></script>
+  <script type= "module" src="./index.js"></script>
+  <script type= "module" src="./modules/time.js"></script>
 </body>
 </html>
